requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap4==4.0.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.26.0
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.1
